+++ isomorphic/index.js
@@ -11,9 +11,6 @@ export function relement__use(relement) {
 	mathml_tags = tagsNS?.('http://www.w3.org/1998/Math/MathML')
 	fragment_ = relement?.fragment_ || '*__fragment__relement'
 	raw_ = relement?.raw_ || '*__fragment__relement'
-	doc_html_ = relement?.doc_html_ || 'server__full__relement'
-	server__element__proto = relement?.server__element__proto || 'server__full__relement'
-	hydrate = relement?.hydrate || 'browser__full__relement'
 }
 export let attach
 export let bind_
@@ -23,9 +20,6 @@ export let svg_tags
 export let mathml_tags
 export let fragment_ = '*__fragment__relement'
 export let raw_ = '*__fragment__relement'
-export let doc_html_ = 'server__full__relement'
-export let server__element__proto = 'server__full__relement'
-export let hydrate = 'browser__full__relement'
 export function svg_tags_() {
 	return svg_tags
 }